BM-845 template to support upload only

+++ app/static/js/directives/UploadManager.js
@@ -2,16 +2,33 @@ BenefitMyApp.directive('bmuploadmanager',
   function() {
     return {
       restrict: 'E',
-      scope: {},
+      scope: {
+        // Call back when an upload is successfully uploaded
+        onUploadAdded: '&',
+        // Call back when and upload is deleted
+        onUploadDeleted: '&',
+        // Template ID to handle
+        templateId: '=',
+        // Document ID to handle
+        documentId: '='
+      },
       templateUrl: '/static/partials/common/upload.html',
       controller: ['$scope',
                    '$timeout',
                    '$attrs',
                    'UploadService',
+                   'TemplateService',
           function($scope,
                    $timeout,
                    $attrs,
-                   UploadService) {
+                   UploadService,
+                   TemplateService) {
+
+            // For the use case where either an existing template ID
+            // or a document ID is given, don't actually delete the 
+            // relevant uploads, as those could be shared with others
+            $scope.hideAndNotDelete = $scope.templateId || $scope.documentId;
+
             $scope.uploadManager = {
               hideUploadArea: false,
               canManageUpload: true,
@@ -22,14 +39,31 @@ BenefitMyApp.directive('bmuploadmanager',
               uploadedFiles: [],
               files:[],
               deleteS3File: function(file){
-                UploadService.deleteFile(file.id, file.S3).then(function(deletedFile){
-                  $scope.uploadManager.uploadedFiles = _.without($scope.uploadManager.uploadedFiles, file);
-                  $scope.uploadManager.deleteSuccess = true;
-                  $timeout(function(){
-                    $scope.uploadManager.deleteSuccess = false;
-                  }, 5000);
-                });
-              }};
+                // If indicated so, do not really delete the files, but rather hide from
+                // from the view. 
+                if ($scope.hideAndNotDelete) {
+                    $scope.uploadManager.uploadedFiles = _.without($scope.uploadManager.uploadedFiles, file);
+                    $scope.uploadManager.deleteSuccess = true;
+                    if ('onUploadDeleted' in $attrs) {
+                        $scope.onUploadDeleted({uploadId: file.id});
+                    }
+                    $timeout(function(){
+                        $scope.uploadManager.deleteSuccess = false;
+                    }, 5000);
+                }
+                else {
+                    UploadService.deleteFile(file.id, file.S3).then(function(deletedFile){
+                      $scope.uploadManager.uploadedFiles = _.without($scope.uploadManager.uploadedFiles, file);
+                      $scope.uploadManager.deleteSuccess = true;
+                      if ('onUploadDeleted' in $attrs) {
+                        $scope.onUploadDeleted({uploadId: file.id});
+                      }
+                      $timeout(function(){
+                        $scope.uploadManager.deleteSuccess = false;
+                      }, 5000);
+                    });
+                }
+              }}; 
 
             var handleUploadArea = function(files, uploadType){
               if (files && files.length) {
@@ -40,6 +74,9 @@ BenefitMyApp.directive('bmuploadmanager',
                     function(fileUploaded){
                       $scope.uploadManager.inProgress = undefined;
                       $scope.uploadManager.uploadedFiles.unshift(fileUploaded);
+                      if ('onUploadAdded' in $attrs) {
+                        $scope.onUploadAdded({uploadId: fileUploaded.id});
+                      }
                       if($attrs.featureId){
                         UploadService.SetUploadApplicationFeature(fileUploaded.id, uploadType, $attrs.featureId)
                         .then(function(){
@@ -58,19 +95,38 @@ BenefitMyApp.directive('bmuploadmanager',
                 }
               }
             };
+
+            $scope.hideUploadArea = function() {
+                return $scope.uploadManager.hideUploadArea 
+                    || ($attrs.maxUploads 
+                        && $scope.uploadManager.uploadedFiles
+                        && $scope.uploadManager.uploadedFiles.length >= $attrs.maxUploads);
+            };
+
             $scope.$watch('uploadManager.files', function(){
               handleUploadArea($scope.uploadManager.files, $attrs.uploadType);
             });
 
-            if($attrs.featureId && $attrs.uploadType){
-              $attrs.$observe('featureId', function(){
+            if($attrs.featureId && $attrs.uploadType) {
+              $attrs.$observe('featureId', function() {
                 UploadService.getUploadsByFeature($attrs.featureId, $attrs.uploadType)
                 .then(function(resp){
                   $scope.uploadManager.uploadedFiles = resp;
                 });
               });
+            } 
+            else if ('templateId' in $attrs) {
+                TemplateService.getTemplateById($scope.templateId)
+                  .then(function(template){
+                    if (template && template.upload) {
+                        $scope.uploadManager.uploadedFiles.push(template.upload);
+                    }
+                  });
             }
-            else{
+            else if ('documentId' in $attrs) {
+
+            }
+            else {
               UploadService.getAllUploadsByCurrentUser().then(function(resp){
                 $scope.uploadManager.uploadedFiles = resp;
               });
